:hammer: refactored to new structure
