+++ src/assets/icons/IconArrowRight.tsx
