+++ src/assets/icons/IconChart.tsx
