+++ src/assets/icons/IconSettings.tsx
